+++ src/main/java/cms/gov/madie/measure/services/PackageService.java
@@ -7,5 +7,7 @@ import gov.cms.madie.models.measure.Measure;
 public interface PackageService {
   PackageDto getMeasurePackage(Measure measure, String accessToken);
 
+  String getHumanReadable(Measure measure, String username, String accessToken);
+
   byte[] getQRDA(QrdaRequestDTO serviceDTO, String accessToken);
 }
