+++ src/main/resources/application.yml
@@ -51,6 +51,7 @@ madie:
     base-url: ${QDM_SERVICE_URL:http://localhost:8086/api}
     create-package-urn: /qdm/measures/package
     create-qrda-urn: /qdm/measures/package/qrda
+    human-readable: /qdm/measures/human-readable
     retrieve-cqmmeasure-urn: /qdm/measures/cqm
   service-config:
     json-url: ${SERVICE_CONFIG_JSON_URL:http://localhost:9000/env-config/serviceConfig.json}
